backend/requirements.txt
fastapi
uvicorn
python-multipart
ffmpeg-python
google-cloud-aiplatform
python-dotenv
google-generativeai
google-cloud-storage
